+++ src/main/java/be/baur/sds/content/BooleanType.java
@@ -1,13 +1,13 @@
 package be.baur.sds.content;
 
-import be.baur.sds.MixedType;
+import be.baur.sds.DataType;
 import be.baur.sds.common.Content;
 
 /**
  * A <code>BooleanType</code> is a simple type representing an SDA node with
  * boolean content,that is "true" or "false". No other values are allowed.
  */
-public final class BooleanType extends MixedType {
+public final class BooleanType extends DataType {
 
 	public static final String TRUE = "true";
 	public static final String FALSE = "false";