+++ src/main/java/be/baur/sds/content/RangedType.java
@@ -2,7 +2,7 @@ package be.baur.sds.content;
 
 import java.util.Objects;
 
-import be.baur.sds.MixedType;
+import be.baur.sds.DataType;
 import be.baur.sds.common.Interval;
 
 /**
@@ -10,7 +10,7 @@ import be.baur.sds.common.Interval;
  * within an interval, like an integer, a decimal or a date(time).<br>
  * See also {@link Interval}.
  */
-public abstract class RangedType <T extends Comparable<?>> extends MixedType {
+public abstract class RangedType <T extends Comparable<?>> extends DataType {
 
 	private Interval<?> range = Interval.MIN_TO_MAX; // default any value is allowed.